Add deps uberjar generation rule
This can replace a custom `genrule`, so you can just call
`babashka_deps_jar` and get a jar file output.

Because we want it to run within a sandbox, it uses a local m2
repository dir instead of falling back to the system one (which it
cannot write to). This is fine, because the process will only be re-run
when the input `bb.edn` file changes.